Fix Focus state
